refactor: rename components to match code style
