+++ apps/client/components/notebook/monaco-editor-client.tsx
