+++ apps/client/components/notebook/outline-panel.tsx
