+++ apps/client/components/notebook/output-view.tsx
