+++ apps/client/components/notebook/setup-panel.tsx
