+++ package.json
@@ -4,7 +4,13 @@
   "private": true,
   "version": "0.1.0",
   "repository": "https://github.com/julianduque/nodebooks",
-  "keywords": ["node", "notebook", "fastify", "nextjs", "typescript"],
+  "keywords": [
+    "node",
+    "notebook",
+    "fastify",
+    "nextjs",
+    "typescript"
+  ],
   "license": "MIT",
   "author": "Julián Duque <julianduquej@gmail.com>",
   "workspaces": [
