+++ packages/notebook-ui/src/index.tsx
@@ -5,16 +5,16 @@ import type { UiDisplay } from "@nodebooks/notebook-schema";
 import { Image } from "./components/image";
 import { Markdown } from "./components/markdown";
 import { HtmlBlock } from "./components/html";
-import { JsonViewer } from "./components/jsonViewer";
-import { CodeBlock } from "./components/codeBlock";
-import { TableGrid } from "./components/tableGrid";
-import { DataSummary } from "./components/dataSummary";
+import { JsonViewer } from "./components/json-viewer";
+import { CodeBlock } from "./components/code-block";
+import { TableGrid } from "./components/table-grid";
+import { DataSummary } from "./components/data-summary";
 import { AlertCallout } from "./components/alert";
 import { BadgeTag } from "./components/badge";
 import { MetricTile } from "./components/metric";
 import { ProgressBar } from "./components/progress";
 import { Spinner } from "./components/spinner";
-import { UiCard } from "./components/uiCard";
+import { UiCard } from "./components/ui-card";
 
 // UI Renderer
 export interface UiRendererProps {
@@ -77,13 +77,13 @@ export default UiRenderer;
 export { Image } from "./components/image";
 export { Markdown } from "./components/markdown";
 export { HtmlBlock } from "./components/html";
-export { JsonViewer } from "./components/jsonViewer";
-export { CodeBlock } from "./components/codeBlock";
-export { TableGrid } from "./components/tableGrid";
-export { DataSummary } from "./components/dataSummary";
+export { JsonViewer } from "./components/json-viewer";
+export { CodeBlock } from "./components/code-block";
+export { TableGrid } from "./components/table-grid";
+export { DataSummary } from "./components/data-summary";
 export { AlertCallout } from "./components/alert";
 export { BadgeTag } from "./components/badge";
 export { MetricTile } from "./components/metric";
 export { ProgressBar } from "./components/progress";
 export { Spinner } from "./components/spinner";
-export { UiCard } from "./components/uiCard";
+export { UiCard } from "./components/ui-card";
